Printer for invoice & Invoice number incrment
For printer, The only change is the new method printer() near the bottom of code and code for createInvoiceScene(). You want to put a button in InvoiceScene to print and use the textfield name for the argument of the printer() method not the textfield.text since printer() takes a node/object rather than a String. I think there is nothing to import if I'm wrong.

For Invoice increment, it is a 2 step process. One is to read the textfile to get current invoice number and then one to overwrite the the textfile. 2 methods on the bottom called readInvoice() and overWriteInvoice(). Both methods are looking for the textfile called invoice.txt. You can just make the textfile your self and call it just that (Invoice.txt) and put it next to where the style.css file is (basically inside the src folder). I put a comment on top of the code file for what needs to be imported.

Change the code on top where the variable invoiceNumber is to (private int invoiceNumber = readInvoiceFile();).
The other is change some of the code in goToInvoiceScene(), at the bottom of each (if (chocie == "")) there is the code (						
						overWriteInvoice(invoiceNumber);
						invoiceNumber++;
)
Also in the String invoice variable I changed the variable invoiceNumber to ( String.format("%010d%n", invoiceNumber) ) to have it show to 10 decimal places to the left of whatever value invoice is to give it the (0000000123) number.
I think that's it.