This is the complete project - CSS style (Import in eclipse as project)
Had to zip since drag and drop folder. Made it upload every file in the folder instead of the folder itself.
Total 7 classes. We started with 2 (GUI and Driver) but now with each module/invoice which adds 5 more there is now 7.

Basically all objects are moved out into its own class for each scene. the methods cannot be moved out of GUI so had to do reference argument calling in GUI.